+++ benchmark/src/main/scala/benchmark/wrapper/ldbc/Select.scala
@@ -19,7 +19,7 @@ import cats.effect.unsafe.implicits.global
 
 import ldbc.sql.*
 
-import ldbc.query.builder.syntax.io.*
+import ldbc.query.builder.syntax.*
 
 import jdbc.connector.*
