+++ docs/src/main/scala/00-Setup.scala
@@ -9,8 +9,7 @@ import cats.syntax.all.*
 import cats.effect.*
 import cats.effect.unsafe.implicits.global
 
-import ldbc.dsl.io.*
-import ldbc.dsl.DBIO
+import ldbc.dsl.*
 
 import ldbc.connector.*
