+++ tests/src/test/scala/ldbc/tests/CodecTest.scala
@@ -16,7 +16,7 @@ import munit.*
 
 import ldbc.sql.*
 
-import ldbc.dsl.io.*
+import ldbc.dsl.*
 
 import ldbc.connector.{ MySQLProvider as LdbcProvider, * }
