+++ tests/src/test/scala/ldbc/tests/DDLTest.scala
@@ -17,7 +17,7 @@ import ldbc.sql.*
 import ldbc.dsl.*
 
 import ldbc.schema.*
-import ldbc.schema.syntax.io.*
+import ldbc.schema.syntax.*
 
 import ldbc.connector.{ MySQLProvider as LdbcProvider, * }
