+++ tests/src/test/scala/ldbc/tests/SQLStringContextQueryTest.scala
@@ -16,8 +16,8 @@ import munit.CatsEffectSuite
 
 import ldbc.sql.*
 
+import ldbc.dsl.*
 import ldbc.dsl.exception.DecodeFailureException
-import ldbc.dsl.io.*
 
 import ldbc.connector.{ MySQLProvider as LdbcProvider, * }
