+++ tests/src/test/scala/ldbc/tests/TableQueryUpdateConnectionTest.scala
@@ -20,7 +20,7 @@ import ldbc.sql.*
 import ldbc.dsl.*
 
 import ldbc.query.builder.*
-import ldbc.query.builder.syntax.io.*
+import ldbc.query.builder.syntax.*
 
 import ldbc.connector.{ MySQLProvider as LdbcProvider, * }
