+++ tests/src/test/scala/ldbc/tests/TableSchemaUpdateConnectionTest.scala
@@ -19,7 +19,7 @@ import ldbc.sql.*
 
 import ldbc.dsl.*
 
-import ldbc.schema.syntax.io.*
+import ldbc.schema.syntax.*
 import ldbc.schema.TableQuery
 
 import ldbc.connector.{ MySQLProvider as LdbcProvider, * }
